Remove lanczos function (keep deprecated alias)
