+++ tests/test_exact_diag.py
@@ -5,7 +5,7 @@ import tenpy.linalg.np_conserved as npc
 import numpy as np
 from tenpy.models.xxz_chain import XXZChain
 from tenpy.algorithms.exact_diag import ExactDiag
-from tenpy.linalg.krylov_based import lanczos
+from tenpy.linalg.krylov_based import LanczosGroundState
 
 
 def test_ED():
@@ -37,7 +37,7 @@ def test_ED():
     # check if we can also do lanczos.
     np.random.seed(12345)
     psi3 = npc.Array.from_func(np.random.random, psi2.legs, qtotal=psi2.qtotal, shape_kw='size')
-    E0, psi3, N = lanczos(ED2, psi3)
+    E0, psi3, N = LanczosGroundState(ED2, psi3, {}).run()
     print("Lanczos E0 =", E0)
     ov = npc.inner(psi3, psi2, 'range', do_conj=True)
     print("overlab <psi2 | psi3> = 1. -", 1. - ov)
